+++ src/app/app-routing.module.ts
@@ -1,19 +1,19 @@
 //app-routing.module.ts
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { SearchCityComponent } from './example/city/search.city.component';
-import { SearchPlantComponent } from './example/plant/search.plant.component';
-import { SearchStateComponent } from './example/us-state/search.state.component';
-import { UsStateComponent } from './example/us-state/us.state.component';
+import { SearchCityComponent } from './sandbox/city/search.city.component';
+import { SearchPlantComponent } from './sandbox/plant/search.plant.component';
+import { SearchStateComponent } from './sandbox/us-state/search.state.component';
+import { UsStateComponent } from './sandbox/us-state/us.state.component';
 import { HomeComponent } from './home/home.component';
 
 const routes: Routes = [
   { path: '', component: HomeComponent },
   { path: 'home', component: HomeComponent },
-  { path: 'example/us-state', component: UsStateComponent },
-  { path: 'example/search-city', component: SearchCityComponent },
-  { path: 'example/search-plant', component: SearchPlantComponent },
-  { path: 'example/search-state', component: SearchStateComponent },
+  { path: 'sandbox/us-state', component: UsStateComponent },
+  { path: 'sandbox/search-city', component: SearchCityComponent },
+  { path: 'sandbox/search-plant', component: SearchPlantComponent },
+  { path: 'sandbox/search-state', component: SearchStateComponent },
   { path: '**', component: HomeComponent }
 ];
 @NgModule({
