+++ src/app/core/navbar-fixed/navbar-fixed.component.html
@@ -19,10 +19,10 @@
           <fa-icon [icon]="faAngleDown"></fa-icon>
         </a>
         <ul>
-          <li><a routerLink="example/search-city">Search City Filter</a></li>
-          <li><a routerLink="example/search-plant">Search Plant Filter</a></li>
-          <li><a routerLink="example/us-state">US States</a></li>
-          <li><a routerLink="example/search-state">Search US States</a></li>
+          <li><a routerLink="sandbox/search-city">Search City Filter</a></li>
+          <li><a routerLink="sandbox/search-plant">Search Plant Filter</a></li>
+          <li><a routerLink="sandbox/us-state">US States</a></li>
+          <li><a routerLink="sandbox/search-state">Search US States</a></li>
         </ul>
       </li>
