+++ src/app/sandbox/plant/search.plant.component.scss
